requirements.txt
pydantic==2.5.3
llama-index>=0.10.0
fastapi
python-dotenv
pdfplumber
qdrant-client
openai
uvicorn